requirements.txt
langchain
langchain-core
langchain-community
langchain-mistralai
python-dotenv
fastapi
uvicorn
pydantic